+++ internal/shared/shared.go
@@ -1,7 +1,15 @@
 package shared
 
-import "strings"
+import (
+	"os"
+	"strings"
+)
 
 func GetStringLines(input string) []string {
 	return strings.Split(input, "\n")
+}
+
+func FileExists(filePath string) bool {
+	_, err := os.Stat(filePath)
+	return !os.IsNotExist(err)
 }
